requirements.txt
streamlit
ib_insync
pyyaml
fastapi
uvicorn
requests
pydantic
nest_asyncio
dash==2.14.0
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.1
requests==2.31.0
pyyaml==6.0.1